moving interfaces around
