+++ protocol/implementation/src/mock-pfi/Agent.ts
@@ -1,4 +1,4 @@
-import { TBDexMessage, RequestForQuote, Quote } from "./TBDexProtocol";
+import { TBDexMessage, RequestForQuote, Quote } from "../interfaces/tbdex-types";
 
 
 export function selectBid(
